Rename to TransileConfig
