+++ webpack.config.js
@@ -18,7 +18,7 @@ const CircularDependencyPlugin = require( 'circular-dependency-plugin' );
 const DuplicatePackageCheckerPlugin = require( 'duplicate-package-checker-webpack-plugin' );
 const MomentTimezoneDataPlugin = require( 'moment-timezone-data-webpack-plugin' );
 const SassConfig = require( '@automattic/calypso-build/webpack/sass' );
-const JsxConfig = require( '@automattic/calypso-build/webpack/jsx' );
+const TranspileConfig = require( '@automattic/calypso-build/webpack/transpile' );
 
 /**
  * Internal dependencies
@@ -180,7 +180,7 @@ function getWebpackConfig( {
 			// https://github.com/localForage/localForage/issues/577
 			noParse: /[\/\\]node_modules[\/\\]localforage[\/\\]dist[\/\\]localforage\.js$/,
 			rules: [
-				JsxConfig.loader( {
+				TranspileConfig.loader( {
 					workerCount,
 					configFile: path.join( __dirname, 'babel.config.js' ),
 					cacheDirectory: path.join( __dirname, 'build', '.babel-client-cache' ),
